Split out components in docs
